// src/types/FNBINFSAL.ts
export interface FNBINFSAL {
  ID?: number;
  FNCPERSON_ID: number;
  PESO?: number;
  TALLA?: number;
  TA_SISTOLICA?: number;
  TA_DIASTOLICA?: number;
  USO_PROTESIS?: number;
  TIEMPO_PROTESIS?: number;
  ULTIMA_VISITA?: Date;
  FECHA_MUERTE?: Date;
  FNCINTIMC_ID?: number;
  FNCINTTEA_ID?: number;
  ESTADO?: number;
}